remove add node modal, update addNewNode func, ut, and snapshot
