+++ src/data/manifests_v.10.json
@@ -1,1 +1,1 @@
-{"_v": 10, "commands": [{"name": "Add", "key": "ADD", "type": "ARITHMETIC", "inputs": [{"name": "y", "id": "add_y", "type": "target"}]}, {"name": "Local File", "key": "LOCAL_FILE", "type": "LOADER"}, {"name": "Butterworth Filter", "key": "BUTTER", "type": "SIGNAL_PROCESSING"}, {"name": "Invert", "key": "INVERT", "type": "ARRAY_AND_MATRIX"}, {"name": "Matmul", "key": "MATMUL", "type": "ARRAY_AND_MATRIX", "inputs": [{"name": "y", "id": "matmul_y", "type": "target"}]}, {"name": "Image Viewer", "key": "IMAGE", "type": "VISOR"}, {"name": "Object Detection", "key": "OBJECT_DETECTION", "type": "SIMULATION"}, {"name": "3D Surface Plot", "key": "SURFACE3D", "type": "VISOR"}, {"name": "Linspace", "key": "LINSPACE", "type": "SIMULATION", "ui_component_id": "default", "end": {"type": "float", "default": "0"}, "step": {"type": "float", "default": "1000"}}, {"name": "3D Scatter Plot", "key": "SCATTER3D", "type": "VISOR", "inputs": []}, {"name": "Savitzky-Golay Filter", "key": "SAVGOL", "type": "SIGNAL_PROCESSING"}, {"name": "Random (Uniform)", "key": "RAND", "type": "SIMULATION"}, {"name": "Multiply", "key": "MULTIPLY", "type": "ARITHMETIC", "inputs": [{"name": "y", "id": "multiply_y", "type": "target"}]}, {"name": "Bar Chart", "key": "BAR", "type": "VISOR"}, {"name": "loop", "key": "LOOP", "type": "LOOP", "inputs": [{"name": "end", "id": "end", "type": "source"}, {"name": "body", "id": "body", "type": "source"}]}, {"name": "Sine Wave", "key": "SINE", "type": "SIMULATION"}, {"name": "Square Wave", "key": "SQUARE", "type": "SIMULATION"}, {"name": "Sawtooth", "key": "SAW", "type": "SIMULATION"}, {"name": "loop-conditional", "key": "CONDITIONAL", "type": "CONDITIONAL"}, {"name": "Histogram", "key": "HISTOGRAM", "type": "VISOR"}, {"name": "Constant", "key": "CONSTANT", "type": "SIMULATION"}, {"name": "Time Series", "key": "TIMESERIES", "type": "SIMULATION"}, {"name": "conditional", "key": "CONDITIONAL", "type": "CONDITIONAL", "inputs": [{"name": "x", "id": "first", "type": "target"}, {"name": "y", "id": "second", "type": "target"}, {"name": "true", "id": "true", "type": "source"}, {"name": "false", "id": "false", "type": "source"}]}, {"name": "Line", "key": "LINE", "type": "VISOR"}, {"name": "Scatter Plot", "key": "SCATTER", "type": "VISOR"}, {"name": "Sub", "key": "SUBTRACT", "type": "ARITHMETIC", "inputs": [{"name": "y", "id": "sub_y", "type": "target"}]}, {"name": "timer", "key": "TIMER", "type": "TIMER"}], "parameters": {"LOCAL_FILE": {"file_type": {"type": "string", "default": "image"}, "op_type": {"type": "string", "default": "OD"}, "path": {"type": "string", "default": ""}}, "INVERT": {"rcond": {"type": "float", "default": 1.0}}, "LINSPACE": {"start": {"type": "float", "default": "10"}, "end": {"type": "float", "default": "0"}, "step": {"type": "float", "default": "1000"}}, "SAVGOL": {"wlen": {"type": "int", "default": 3}, "porder": {"type": "int", "default": 1}}, "LOOP": {"initial_count": {"type": "int", "default": 0}, "iteration_count": {"type": "int", "default": 2}, "step": {"type": "int", "default": 1}}, "SINE": {"frequency": {"type": "float", "default": 1}, "offset": {"type": "float", "default": 0}, "amplitude": {"type": "float", "default": 1}, "waveform": {"type": "select", "options": ["sine", "square", "triangle", "sawtooth"], "default": "sine"}}, "CONDITIONAL": {"operator_type": {"type": "string", "default": ">="}}, "CONSTANT": {"constant": {"type": "float", "default": "3"}}, "TIMESERIES": {}, "TIMER": {"sleep_time": {"type": "int", "default": 0}}}}
+{"_v": 10, "commands": [{"name": "conditional", "key": "CONDITIONAL", "type": "CONDITIONAL", "inputs": [{"name": "x", "id": "first", "type": "target"}, {"name": "y", "id": "second", "type": "target"}, {"name": "true", "id": "true", "type": "source"}, {"name": "false", "id": "false", "type": "source"}]}, {"name": "Invert", "key": "INVERT", "type": "ARRAY_AND_MATRIX"}, {"name": "loop", "key": "LOOP", "type": "LOOP", "inputs": [{"name": "end", "id": "end", "type": "source"}, {"name": "body", "id": "body", "type": "source"}]}, {"name": "timer", "key": "TIMER", "type": "TIMER"}, {"name": "Sine Wave", "key": "SINE", "type": "SIMULATION"}, {"name": "Square Wave", "key": "SQUARE", "type": "SIMULATION"}, {"name": "Sawtooth", "key": "SAW", "type": "SIMULATION"}, {"name": "3D Scatter Plot", "key": "SCATTER3D", "type": "VISOR", "inputs": []}, {"name": "Multiply", "key": "MULTIPLY", "type": "ARITHMETIC", "inputs": [{"name": "y", "id": "multiply_y", "type": "target"}]}, {"name": "Feedback", "key": "FEEDBACK", "type": "SIMULATION"}, {"name": "Time Series", "key": "TIMESERIES", "type": "SIMULATION"}, {"name": "Scatter Plot", "key": "SCATTER", "type": "VISOR"}, {"name": "Image Viewer", "key": "IMAGE", "type": "VISOR"}, {"name": "Local File", "key": "LOCAL_FILE", "type": "LOADER"}, {"name": "Object Detection", "key": "OBJECT_DETECTION", "type": "SIMULATION"}, {"name": "Butterworth Filter", "key": "BUTTER", "type": "SIGNAL_PROCESSING"}, {"name": "Linspace", "key": "LINSPACE", "type": "SIMULATION", "ui_component_id": "default", "end": {"type": "float", "default": "0"}, "step": {"type": "float", "default": "1000"}}, {"name": "Random (Uniform)", "key": "RAND", "type": "SIMULATION"}, {"name": "Add", "key": "ADD", "type": "ARITHMETIC", "inputs": [{"name": "y", "id": "add_y", "type": "target"}]}, {"name": "Histogram", "key": "HISTOGRAM", "type": "VISOR"}, {"name": "Low-pass FIR Filter", "key": "FIR", "type": "SIGNAL_PROCESSING"}, {"name": "3D Surface Plot", "key": "SURFACE3D", "type": "VISOR"}, {"name": "GoTo", "key": "GOTO", "type": "LOOP"}, {"name": "Line", "key": "LINE", "type": "VISOR"}, {"name": "Sub", "key": "SUBTRACT", "type": "ARITHMETIC", "inputs": [{"name": "y", "id": "sub_y", "type": "target"}]}, {"name": "Matmul", "key": "MATMUL", "type": "ARRAY_AND_MATRIX", "inputs": [{"name": "y", "id": "matmul_y", "type": "target"}]}, {"name": "Bar Chart", "key": "BAR", "type": "VISOR"}, {"name": "Savitzky-Golay Filter", "key": "SAVGOL", "type": "SIGNAL_PROCESSING"}, {"name": "Constant", "key": "CONSTANT", "type": "SIMULATION"}, {"name": "The End", "key": "END", "type": "TERMINATOR"}], "parameters": {"CONDITIONAL": {"operator_type": {"type": "string", "default": ">="}}, "INVERT": {"rcond": {"type": "float", "default": 1.0}}, "LOOP": {"num_loops": {"type": "int", "default": -1}}, "TIMER": {"sleep_time": {"type": "int", "default": 0}}, "SINE": {"frequency": {"type": "float", "default": 1}, "offset": {"type": "float", "default": 0}, "amplitude": {"type": "float", "default": 1}, "phase": {"type": "float", "default": 0.0}, "waveform": {"type": "select", "options": ["sine", "square", "triangle", "sawtooth"], "default": "sine"}}, "FEEDBACK": {"referred_node": {"type": "node_reference", "default": ""}}, "TIMESERIES": {}, "LOCAL_FILE": {"file_type": {"type": "string", "default": "image"}, "op_type": {"type": "string", "default": "OD"}, "path": {"type": "string", "default": ""}}, "LINSPACE": {"start": {"type": "float", "default": "10"}, "end": {"type": "float", "default": "0"}, "step": {"type": "float", "default": "1000"}}, "FIR": {"sample_rate": {"type": "float", "default": 100}, "transition_width": {"type": "float", "default": 5.0}, "stop_band_attenuation": {"type": "float", "default": 60}, "cutoff_freq": {"type": "float", "default": 10.0}}, "GOTO": {"goto_node_id": {"type": "node_reference", "default": ""}}, "SAVGOL": {"wlen": {"type": "int", "default": 3}, "porder": {"type": "int", "default": 1}}, "CONSTANT": {"constant": {"type": "float", "default": "3"}}, "END": {}}}
