+++ src/feature/flow_chart_panel/views/ControlBar.tsx
@@ -1,6 +1,5 @@
-import { memo, useCallback, FC, useState, useEffect, useRef } from "react";
+import { memo, FC, useState, useEffect } from "react";
 import localforage from "localforage";
-import { v4 as uuidv4 } from "uuid";
 import "react-tabs/style/react-tabs.css";
 import {
   saveFlowChartToLocalStorage,
@@ -9,28 +8,18 @@ import {
 } from "@src/services/FlowChartServices";
 import { useFlowChartState } from "@src/hooks/useFlowChartState";
 import ReactSwitch from "react-switch";
-import AddNodeModal from "./AddNodeModal";
 import { ControlsProps } from "../types/ControlsProps";
-import { NodeOnAddFunc, ParamTypes } from "../types/NodeAddFunc";
 import { useSocket } from "@src/hooks/useSocket";
 import CancelIconSvg from "@src/utils/cancel_icon";
 import PlayBtn from "../components/play-btn/PlayBtn";
 import { IServerStatus } from "@src/context/socket.context";
 import DropDown from "@src/feature/common/dropdown/DropDown";
 import KeyboardShortcutModal from "./KeyboardShortcutModal";
-import { ElementsData } from "../types/CustomNodeProps";
 
 localforage.config({
   name: "react-flow",
   storeName: "flows",
 });
-
-const getNodePosition = () => {
-  return {
-    x: 50 + Math.random() * 20,
-    y: 50 + Math.random() + Math.random() * 20,
-  };
-};
 
 const Controls: FC<ControlsProps> = ({
   theme,
@@ -39,18 +28,10 @@ const Controls: FC<ControlsProps> = ({
 }) => {
   const { states } = useSocket();
   const { socketId, setProgramResults, serverStatus } = states!;
-  const [modalIsOpen, setIsOpen] = useState(false);
   const [isKeyboardShortcutOpen, setIskeyboardShortcutOpen] = useState(false);
 
-  const {
-    isEditMode,
-    setIsEditMode,
-    rfInstance,
-    openFileSelector,
-    saveFile,
-    nodes,
-    setNodes,
-  } = useFlowChartState();
+  const { isEditMode, setIsEditMode, rfInstance, openFileSelector, saveFile } =
+    useFlowChartState();
   const onSave = async () => {
     if (rfInstance && rfInstance.nodes.length > 0) {
       saveFlowChartToLocalStorage(rfInstance);
@@ -69,62 +50,6 @@ const Controls: FC<ControlsProps> = ({
     } else {
       alert("There is no running job on server.");
     }
-  };
-
-  const addNewNode: NodeOnAddFunc = useCallback(
-    ({ funcName, params, type, inputs, uiComponentId }) => {
-      let nodeLabel: string;
-      const nodeId = `${funcName}-${uuidv4()}`;
-      if (funcName === "CONSTANT") {
-        nodeLabel = "2.0";
-      } else {
-        const numOfThisNodesOnChart = nodes.filter(
-          (node) => node.data.func === funcName
-        ).length;
-        nodeLabel = numOfThisNodesOnChart > 0 ? `${funcName}_${numOfThisNodesOnChart}` : funcName;
-      }
-      const nodeParams = params
-        ? Object.keys(params).reduce(
-            (prev: ElementsData["ctrls"], param) => ({
-              ...prev,
-              [param]: {
-                functionName: funcName,
-                param,
-                value:
-                  funcName === "CONSTANT"
-                    ? nodeLabel
-                    : params![param].default?.toString(),
-              },
-            }),
-            {}
-          )
-        : {};
-
-      const newNode = {
-        id: nodeId,
-        type: uiComponentId || type,
-        data: {
-          id: nodeId,
-          label: nodeLabel,
-          func: funcName,
-          type,
-          ctrls: nodeParams,
-          inputs,
-        },
-        position: getNodePosition(),
-      };
-      setNodes((els) => els.concat(newNode));
-      closeModal();
-    },
-    [nodes,setNodes]
-  );
-
-  const openModal = () => {
-    setIsOpen(true);
-  };
-  const afterOpenModal = () => null;
-  const closeModal = () => {
-    setIsOpen(false);
   };
 
   useEffect(() => {
@@ -154,25 +79,14 @@ const Controls: FC<ControlsProps> = ({
           <span>Cancel</span>
         </button>
       )}
-      {activeTab !== "debug" && activeTab === "visual" ? (
+      {isEditMode && activeTab === "panel" && (
         <AddBtn
-          handleClick={() => {
-            openModal();
-          }}
+          testId={"add-ctrl"}
           theme={theme}
-          testId={"add-node"}
+          handleClick={() => {
+            setOpenCtrlModal((prev) => !prev);
+          }}
         />
-      ) : (
-        isEditMode &&
-        activeTab === "panel" && (
-          <AddBtn
-            testId={"add-ctrl"}
-            theme={theme}
-            handleClick={() => {
-              setOpenCtrlModal((prev) => !prev);
-            }}
-          />
-        )
       )}
       {activeTab !== "debug" && (
         <DropDown
@@ -250,26 +164,20 @@ const Controls: FC<ControlsProps> = ({
         onClose={() => setIskeyboardShortcutOpen(false)}
         theme={theme}
       />
-
-      <AddNodeModal
-        afterOpenModal={afterOpenModal}
-        closeModal={closeModal}
-        modalIsOpen={modalIsOpen}
-        onAdd={addNewNode}
-        theme={theme}
-      />
     </div>
   );
 };
 
 export default memo(Controls);
 
-const AddBtn = ({ handleClick, theme, testId, }) => {
+const AddBtn = ({ handleClick, theme, testId }) => {
   return (
     <button
-    data-cy={testId}
-    data-testid={testId}
-    className="save__controls_button btn__add" onClick={handleClick}>
+      data-cy={testId}
+      data-testid={testId}
+      className="save__controls_button btn__add"
+      onClick={handleClick}
+    >
       {" "}
       <div
         style={{
@@ -283,7 +191,6 @@ const AddBtn = ({ handleClick, theme, testId, }) => {
         style={{
           color: theme === "dark" ? "#fff" : "#000",
         }}
-
       >
         Add
       </div>
